+++ tdom/processor.py
@@ -1,4 +1,3 @@
-import sys
 from typing import cast, Protocol
 from collections.abc import Iterable, Sequence, Callable, Generator
 from functools import lru_cache
@@ -10,14 +9,6 @@ from markupsafe import Markup
 from .callables import get_callable_info, CallableInfo
 from .format import format_interpolation as base_format_interpolation
 from .format import format_template
-from .nodes import (
-    Comment,
-    DocumentType,
-    Element,
-    Fragment,
-    Node,
-    Text,
-)
 from .htmlspec import (
     VOID_ELEMENTS,
     CDATA_CONTENT_ELEMENTS,
@@ -40,21 +31,13 @@ from .parser import (
     TTemplatedAttribute,
     TText,
 )
-from .placeholders import TemplateRef
-from .template_utils import template_from_parts
 from .utils import CachableTemplate, LastUpdatedOrderedDict
-from .protocols import HasHTMLDunder
 from .escaping import (
     escape_html_script as default_escape_html_script,
     escape_html_style as default_escape_html_style,
     escape_html_text as default_escape_html_text,
     escape_html_comment as default_escape_html_comment,
 )
-
-
-@lru_cache(maxsize=0 if "pytest" in sys.modules else 512)
-def _parse_and_cache(cachable: CachableTemplate) -> TNode:
-    return TemplateParser.parse(cachable.template)
 
 
 type Attribute = tuple[str, object]
@@ -329,7 +312,7 @@ def _resolve_t_attrs(
                 else:
                     new_attrs[name] = attr_value
             case TTemplatedAttribute(name=name, value_ref=ref):
-                attr_t = _resolve_ref(ref, interpolations)
+                attr_t = ref.resolve(interpolations)
                 attr_value = format_template(attr_t)
                 if name in ATTR_ACCUMULATOR_MAKERS:
                     if name not in attr_accs:
@@ -377,76 +360,12 @@ def _resolve_html_attrs(attrs: AttributesDict) -> HTMLAttributesDict:
     return html_attrs
 
 
-def _resolve_attrs(
-    attrs: Sequence[TAttribute], interpolations: tuple[Interpolation, ...]
-) -> HTMLAttributesDict:
-    """
-    Substitute placeholders in attributes for HTML elements.
-
-    This is the full pipeline: interpolation + HTML processing.
-    """
-    interpolated_attrs = _resolve_t_attrs(attrs, interpolations)
-    return _resolve_html_attrs(interpolated_attrs)
-
-
-def _flatten_nodes(nodes: Iterable[Node]) -> list[Node]:
-    """Flatten a list of Nodes, expanding any Fragments."""
-    flat: list[Node] = []
-    for node in nodes:
-        if isinstance(node, Fragment):
-            flat.extend(node.children)
-        else:
-            flat.append(node)
-    return flat
-
-
-def _substitute_and_flatten_children(
-    children: Iterable[TNode], interpolations: tuple[Interpolation, ...]
-) -> list[Node]:
-    """Substitute placeholders in a list of children and flatten any fragments."""
-    resolved = [_resolve_t_node(child, interpolations) for child in children]
-    flat = _flatten_nodes(resolved)
-    return flat
-
-
-def _node_from_value(value: object) -> Node:
-    """
-    Convert an arbitrary value to a Node.
-
-    This is the primary action performed when replacing interpolations in child
-    content positions.
-    """
-    match value:
-        case str():
-            return Text(value)
-        case Node():
-            return value
-        case Template():
-            return to_node(value)
-        # Consider: falsey values, not just False and None?
-        case False | None:
-            return Fragment(children=[])
-        case Iterable():
-            children = [_node_from_value(v) for v in value]
-            return Fragment(children=children)
-        case HasHTMLDunder():
-            # CONSIDER: should we do this lazily?
-            return Text(Markup(value.__html__()))
-        case c if callable(c):
-            # Treat all callable values in child content positions as if
-            # they are zero-arg functions that return a value to be processed.
-            return _node_from_value(c())
-        case _:
-            # CONSIDER: should we do this lazily?
-            return Text(str(value))
-
-
 def _kebab_to_snake(name: str) -> str:
     """Convert a kebab-case name to snake_case."""
     return name.replace("-", "_").lower()
 
 
-def _prep_component_kwargs(
+def prep_component_kwargs(
     callable_info: CallableInfo,
     attrs: AttributesDict,
     system_kwargs: dict[str, object],
@@ -476,134 +395,6 @@ def _prep_component_kwargs(
         )
 
     return kwargs
-
-
-def _invoke_component(
-    attrs: AttributesDict,
-    children: list[Node],  # TODO: why not TNode, though?
-    interpolation: Interpolation,
-) -> Node:
-    """
-    Invoke a component callable with the provided attributes and children.
-
-    Components are any callable that meets the required calling signature.
-    Typically, that's a function, but it could also be the constructor or
-    __call__() method for a class; dataclass constructors match our expected
-    invocation style.
-
-    We validate the callable's signature and invoke it with keyword-only
-    arguments, then convert the result to a Node.
-
-    Component invocation rules:
-
-    1. All arguments are passed as keywords only. Components cannot require
-    positional arguments.
-
-    2. Children are passed via a "children" parameter when:
-
-    - Child content exists in the template AND
-    - The callable accepts "children" OR has **kwargs
-
-    If no children exist but the callable accepts "children", we pass an
-    empty tuple.
-
-    3. All other attributes are converted from kebab-case to snake_case
-    and passed as keyword arguments if the callable accepts them (or has
-    **kwargs). Attributes that don't match parameters are silently ignored.
-    """
-    value = format_interpolation(interpolation)
-    if not callable(value):
-        raise TypeError(
-            f"Expected a callable for component invocation, got {type(value).__name__}"
-        )
-    callable_info = get_callable_info(value)
-
-    kwargs = _prep_component_kwargs(
-        callable_info, attrs, system_kwargs={"children": tuple(children)}
-    )
-
-    result = value(**kwargs)
-    return _node_from_value(result)
-
-
-def _resolve_ref(
-    ref: TemplateRef, interpolations: tuple[Interpolation, ...]
-) -> Template:
-    resolved = [interpolations[i_index] for i_index in ref.i_indexes]
-    return template_from_parts(ref.strings, resolved)
-
-
-def _resolve_t_text_ref(
-    ref: TemplateRef, interpolations: tuple[Interpolation, ...]
-) -> Text | Fragment:
-    """Resolve a TText ref into Text or Fragment by processing interpolations."""
-    if ref.is_literal:
-        return Text(ref.strings[0])
-
-    parts = [
-        Text(part)
-        if isinstance(part, str)
-        else _node_from_value(format_interpolation(part))
-        for part in _resolve_ref(ref, interpolations)
-    ]
-    flat = _flatten_nodes(parts)
-
-    if len(flat) == 1 and isinstance(flat[0], Text):
-        return flat[0]
-
-    return Fragment(children=flat)
-
-
-def _resolve_t_node(t_node: TNode, interpolations: tuple[Interpolation, ...]) -> Node:
-    """Resolve a TNode tree into a Node tree by processing interpolations."""
-    match t_node:
-        case TText(ref=ref):
-            return _resolve_t_text_ref(ref, interpolations)
-        case TComment(ref=ref):
-            comment_t = _resolve_ref(ref, interpolations)
-            comment = format_template(comment_t)
-            return Comment(comment)
-        case TDocumentType(text=text):
-            return DocumentType(text)
-        case TFragment(children=children):
-            resolved_children = _substitute_and_flatten_children(
-                children, interpolations
-            )
-            return Fragment(children=resolved_children)
-        case TElement(tag=tag, attrs=attrs, children=children):
-            resolved_attrs = _resolve_attrs(attrs, interpolations)
-            resolved_children = _substitute_and_flatten_children(
-                children, interpolations
-            )
-            return Element(tag=tag, attrs=resolved_attrs, children=resolved_children)
-        case TComponent(
-            start_i_index=start_i_index,
-            end_i_index=end_i_index,
-            attrs=t_attrs,
-            children=children,
-        ):
-            start_interpolation = interpolations[start_i_index]
-            end_interpolation = (
-                None if end_i_index is None else interpolations[end_i_index]
-            )
-            resolved_attrs = _resolve_t_attrs(t_attrs, interpolations)
-            resolved_children = _substitute_and_flatten_children(
-                children, interpolations
-            )
-            # HERE ALSO BE DRAGONS: validate matching start/end callables, since
-            # the underlying TemplateParser cannot do that for us.
-            if (
-                end_interpolation is not None
-                and end_interpolation.value != start_interpolation.value
-            ):
-                raise TypeError("Mismatched component start and end callables.")
-            return _invoke_component(
-                attrs=resolved_attrs,
-                children=resolved_children,
-                interpolation=start_interpolation,
-            )
-        case _:
-            raise ValueError(f"Unknown TNode type: {type(t_node).__name__}")
 
 
 @dataclass
@@ -731,7 +522,7 @@ def interpolate_component(
     if not callable(component_callable):
         raise TypeError("Component callable must be callable.")
 
-    kwargs = _prep_component_kwargs(
+    kwargs = prep_component_kwargs(
         get_callable_info(component_callable),
         _resolve_t_attrs(attrs, template.interpolations),
         system_kwargs=system_kwargs,
@@ -1412,14 +1203,3 @@ _default_process_api = cached_process_service_factory(
 def to_html(template: Template, last_parent_tag: str | None = None) -> str:
     """Parse an HTML t-string, substitue values, and return a string of HTML."""
     return _default_process_api.process_template(template, last_parent_tag)
-
-
-def to_node(template: Template) -> Node:
-    """Parse an HTML t-string, substitue values, and return a tree of Nodes."""
-    cachable = CachableTemplate(template)
-    t_node = _parse_and_cache(cachable)
-    return _resolve_t_node(t_node, template.interpolations)
-
-
-# BWC: SHIM
-html = to_node
